refactor: useThirdPartyAuth
