+++ src/pages/_app.tsx
@@ -13,7 +13,7 @@ import '../shared/ui/toast/toast.css'
 import 'nprogress/nprogress.css'
 import '../shared/ui/date-picker/DatePicker.scss'
 import { StoreProvider } from '@/providers/store-provider/store-provider'
-import { useLoader } from '@/shared/hooks/use-loader'
+import { useLoader } from '@/shared/hooks/useLoader'
 import { ToastNotify } from '@/shared/ui/toast/toast'
 
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
